* Basic Opcode parsing * Basic Parameter Parsing * Handle labels without : * Basic value handling * Cycle Count in listing file * Fixed values shown in listing file

+++ Tools/ZASM/SymbolTable.cs
@@ -39,7 +39,7 @@ namespace ZASM
         public string Name;
         public SymbolType Type;
         public SymbolState State;
-        public int Value;
+        public short Value;
 
         public LineInformation DefiendLine;
